+++ src/Probability/Distribution/Discrete/Blank.php
@@ -1,0 +1,7 @@
+<?php
+namespace Math\Probability\Distribution\Discrete;
+
+class Binomial extends Discrete
+{
+
+}
